pyproject.toml
[project]
name = "poc-crewai-sqlagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=0.121.0",
    "crewai-tools>=0.0.11",
    "sqlalchemy>=2.0.41",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0"
]
